refactor - make new class NumberExtractor
- 입력값 중 숫자 콜렉션을 골라내는 일이 계산기에서 할 일이 아닌 것 같아 따로 클래스로 추출했습니다
