+++ src/test/java/calculator/StringCalculatorTests.java
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -41,32 +39,6 @@ class StringCalculatorTests {
     void addHappyPath(String input) {
         int result = stringCalculator.add(input);
         assertThat(result).isEqualTo(767);
-    }
-
-    @DisplayName("입력값 중 기본 구분자(:와 쉼표)를 제외한 숫자만 골라낸 콜렉션 출력")
-    @ParameterizedTest
-    @ValueSource(strings = {"112:3:5,523:123,1", "112:3:5:523:123:1", "112,3,5,523,123,1"})
-    void compositeParser(String input) {
-        List<Integer> numbers = stringCalculator.extractNumbers(input);
-        assertThat(numbers.get(0)).isEqualTo(112);
-        assertThat(numbers.get(3)).isEqualTo(523);
-    }
-
-    @DisplayName("커스텀 구분자가 있는 경우 커스텀 구분자 반환")
-    @ParameterizedTest
-    @ValueSource(strings = {"//!\n112!3:5,523:123!1"})
-    void getCustomDivider(String input) {
-        String customDivider = stringCalculator.getCustomDivider(input);
-        assertThat(customDivider).isEqualTo("!");
-    }
-
-    @DisplayName("커스텀 구분자가 있는 경우 숫자 콜렉션 추출")
-    @ParameterizedTest
-    @ValueSource(strings = {"//!\n112!3:5,523:123!1"})
-    void getNumbersWhenHasCustomDivider(String input) {
-        List<Integer> numbers = stringCalculator.extractNumbers(input);
-        assertThat(numbers.get(0)).isEqualTo(112);
-        assertThat(numbers.get(3)).isEqualTo(523);
     }
 
     @DisplayName("커스텀 구분자가 포함된 정상적인 문자열이 입력됐을 때 추출된 숫자 콜렉션의 총합을 계산")
